+++ examples/data-sources/ise_trustsec_egress_matrix_cell_default/data-source.tf
@@ -1,0 +1,3 @@
+data "ise_trustsec_egress_matrix_cell_default" "example" {
+  id = "92c1a900-8c01-11e6-996c-525400b48521"
+}