+++ examples/resources/ise_trustsec_egress_matrix_cell_default/import.sh
@@ -1,0 +1,1 @@
+terraform import ise_trustsec_egress_matrix_cell_default.example "92c1a900-8c01-11e6-996c-525400b48521"